Resolved Compiler issue on Linux, with a fix that should also work for Windows.

+++ db/vlog_files_test.cc
@@ -499,7 +499,7 @@ TEST_F(DBVLogTest, VlogFileSizeTest) {
     printf(" %zd",nkeys);
     DestroyAndReopen(options);
     // write the kvs
-    for(int key = 0;key<nkeys;++key){
+    for(int key = 0;key<(int)nkeys;++key){
       ASSERT_OK(Put(Key(key), RandomString(&rnd, value_size)));
     }
     size_t vsize=nkeys*(value_size+5);  // bytes written to vlog.  5 bytes for compression header
